Altered diagram, toString of Printer and Computer and comments
I have switched the composition of the UML diagram to the Computer and
Printer instead of the Store.

I also made some changes to the toString method of Computer and Printer,
I used the super toString of the parent class Device, and concatenated
the attribute value of the Computer or Printer to the end of the toString.