Bootcamp 3 requirements implemented in one night before due date.
I first began by converting the Device class into an abstract class. I
created the method "generateID()". This method was also implemented into
the child classes of Device, there was extra functionality added to it,
from the Utils class. The Utils class contains a static method "nextID()",
this method generates a random integer in the bounds specified in it's
parameters.

I moved onto the Store class to change the arrays "computers" and
"printers" to ArrayLists. This required some work since some of the code
had to be altered to fit the syntax of the ArrayList class. I altered
createComputers() and createPrinters() methods to match the requirements.
The user must now enter the values themselves. This is done by using a
scanner to retrieve user input. The runBazar() method needed the
createComputers() and createPrinters() methods to be removed, so that
the user could use the consoleIO which was implemented next.

The consoleIO was implemented in the main method of the BazarDriver
class. This required the use of the MenuInput classes' static method
menuItem(). This static method takes user input with a scanner and
returns it as an integer. It also displays the text based interface.

I started working on changing the UML diagram to fit the new changes in
the code. I was not able to work on my UML diagram beforehand in class.